gh-113299: Create libclinic package
* Create Tools/clinic/libclinic/ package.
* Move cpp.py to libclinic.
* Create libclinic.utils.
