+++ Tools/clinic/clinic.py
@@ -13,7 +13,6 @@ import builtins as bltns
 import collections
 import contextlib
 import copy
-import cpp
 import dataclasses as dc
 import enum
 import functools
@@ -47,8 +46,12 @@ from typing import (
     Protocol,
     TypeVar,
     cast,
-    overload,
 )
+
+from libclinic import cpp
+from libclinic import utils
+from libclinic.utils import ClinicError, warn, fail
+
 
 # TODO:
 #
@@ -142,77 +145,6 @@ def text_accumulator() -> TextAccumulator:
     """
     text, append, output = _text_accumulator()
     return TextAccumulator(append, output)
-
-
-@dc.dataclass
-class ClinicError(Exception):
-    message: str
-    _: dc.KW_ONLY
-    lineno: int | None = None
-    filename: str | None = None
-
-    def __post_init__(self) -> None:
-        super().__init__(self.message)
-
-    def report(self, *, warn_only: bool = False) -> str:
-        msg = "Warning" if warn_only else "Error"
-        if self.filename is not None:
-            msg += f" in file {self.filename!r}"
-        if self.lineno is not None:
-            msg += f" on line {self.lineno}"
-        msg += ":\n"
-        msg += f"{self.message}\n"
-        return msg
-
-
-@overload
-def warn_or_fail(
-    *args: object,
-    fail: Literal[True],
-    filename: str | None = None,
-    line_number: int | None = None,
-) -> NoReturn: ...
-
-@overload
-def warn_or_fail(
-    *args: object,
-    fail: Literal[False] = False,
-    filename: str | None = None,
-    line_number: int | None = None,
-) -> None: ...
-
-def warn_or_fail(
-    *args: object,
-    fail: bool = False,
-    filename: str | None = None,
-    line_number: int | None = None,
-) -> None:
-    joined = " ".join([str(a) for a in args])
-    if clinic:
-        if filename is None:
-            filename = clinic.filename
-        if getattr(clinic, 'block_parser', None) and (line_number is None):
-            line_number = clinic.block_parser.line_number
-    error = ClinicError(joined, filename=filename, lineno=line_number)
-    if fail:
-        raise error
-    else:
-        print(error.report(warn_only=True))
-
-
-def warn(
-    *args: object,
-    filename: str | None = None,
-    line_number: int | None = None,
-) -> None:
-    return warn_or_fail(*args, filename=filename, line_number=line_number, fail=False)
-
-def fail(
-    *args: object,
-    filename: str | None = None,
-    line_number: int | None = None,
-) -> NoReturn:
-    warn_or_fail(*args, filename=filename, line_number=line_number, fail=True)
 
 
 def quoted_for_c_string(s: str) -> str:
@@ -2681,6 +2613,7 @@ impl_definition block
 
         global clinic
         clinic = self
+        utils.clinic = self
 
     def add_include(self, name: str, reason: str,
                     *, condition: str | None = None) -> None:
